requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==4.2.27
django-allauth==65.13.1
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
reportlab==4.4.7
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.2
whitenoise==6.11.0
